ISPN-5384 ReadCommittedEntry.isNull() always returns false
Fixing this allows us to simplify the GetAllCommand handling.
Additionally, unify tx and non-tx distribution interceptor code.
